+++ config-base-next.js
@@ -4,6 +4,7 @@ module.exports = {
     {
       files: ['**/*.ts', '**/*.tsx'],
       rules: {
+        '@typescript-eslint/no-floating-promises': ['error', { ignoreIIFE: true }],
         '@typescript-eslint/promise-function-async': [
           'error',
           {
@@ -18,6 +19,18 @@ module.exports = {
         '@typescript-eslint/return-await': 'error',
 
         'unicorn/no-null': 'warn',
+      },
+    },
+    {
+      files: ['**/test/**/*'],
+      rules: {
+        '@typescript-eslint/consistent-type-assertions': 'off',
+        '@typescript-eslint/explicit-function-return-type': 'off',
+        '@typescript-eslint/explicit-module-boundary-types': 'off',
+        '@typescript-eslint/no-empty-function': 'off',
+        '@typescript-eslint/require-await': 'off',
+        'unicorn/no-null': 'off',
+        'unicorn/no-useless-undefined': 'off',
       },
     },
   ],